Move modal to Apps component
